指摘取り込み
