+++ src/component/Contents/style.tsx
@@ -10,4 +10,6 @@ export const Container = styled.div`
   align-content: flex-start;
   align-items: flex-start;
   justify-content: flex-start;
+  flex: grip;
+  gap: 5% 10%;
 `;
